Bad buggy..

+++ Going Zero/filter/AutoLooper.h
@@ -21,7 +21,7 @@ NS_ASSUME_NONNULL_BEGIN
     SInt32 _currentFrameInLoop;
     SInt32 _loopLengthFrame;
     UInt32 _baseDivider;
-    UInt32 _divider;
+    float _divider;
     float _beatDurationSecForCurrentLoopSession;
     
     UInt32 _state;
